django_server/requirements.txt
django==5.1.2
djangorestframework==3.15.2
django-cors-headers==4.5.0
psycopg2-binary==2.9.10
pillow==11.0.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
django-filter==24.3
whitenoise==6.6.0
django-debug-toolbar==4.4.6
boto3==1.35.73
django-storages==1.14.4
django-cloudinary-storage==0.3.0
cloudinary==1.41.0
gunicorn==23.0.0